restructure the widget classes
