+++ src/Views/shared/header.blade.php
@@ -1,3 +1,3 @@
 @include('ticketit::shared.assets')
-@widget('\Kordy\Ticketit\Widgets\Nav', ['uid' => auth()->user()->id, 'admin_route' => $setting->grab('admin_route')])
+@widget('\Kordy\Ticketit\Views\shared\Nav', ['uid' => auth()->user()->id, 'admin_route' => $setting->grab('admin_route')])
 @include('ticketit::shared.errors')
