+++ src/Views/tickets/show.blade.php
@@ -43,5 +43,6 @@
             });
         });
     </script>
-    @widget('\Kordy\Ticketit\Widgets\SummerNote')
+    @widget('\Kordy\Ticketit\Views\widgets\SummerNote')
+
 @append
