+++ src/Views/widgets/summerNoteTpl.blade.php
